% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Construct cell similarity matrix through expression matrix}
\usage{
calculate_distance(data, method = "pearson", c = 1, nrep = 50)
}
\arguments{
\item{data}{A matrix, rows are genes and columns are cells}

\item{method}{Construct cell similarity matrix through "pearson","euclidean","simlr","rafsil1","rafsil2","phi_s","rho_p"}

\item{c}{Default 1,The number of clusters in the data can only be used for SIMLR}

\item{nrep}{integer for RAFSIL1 the number of times similarities are learned}
}
\value{
Cell similarity matrix
}
\description{
Construct cell similarity matrix through expression matrix
}
\examples{
calculate_distance(data,"pearson")
calculate_distance(data,"euclidean")
calculate_distance(data,"simlr",c=7)
calculate_distance(data,"rafsil1",nrep=50)
calculate_distance(data,"rafsil2")
calculate_distance(data,"phi_s")
calculate_distance(data,"phi_s")
}
