% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{Clustering}
\alias{Clustering}
\title{Use cell similarity matrix clustering to get cluster labels}
\usage{
Clustering(data, c = 1, method = "louvain", iter.max = 1e+09, nstart = 1000)
}
\arguments{
\item{data}{a Matrix}

\item{c}{The number of evaluation clusters of the data}

\item{method}{should "louvain","SC","HC","Kmeans",Respectively represent louvian clustering, spectral clustering, hierarchical clustering, Kmeans clustering}

\item{iter.max}{method="Kmeans":the maximum number of iterations allowed.}

\item{nstart}{method="Kmeans":if centers is a number, how many random sets should be chosen?}
}
\value{
a Vector of cluster label
}
\description{
Use cell similarity matrix clustering to get cluster labels
}
\examples{
Clustering(data,method = "louvain")
Clustering(data,c=7,method = "SC")
Clustering(data,c=7,method = "HC")
Clustering(data,c=7,method = "Kmeans",iter.max = 1e+9, nstart = 1000)
}
