\name{gene_Hypothalamus}
\alias{gene_Hypothalamus}
\docType{data}
\title{A demo:gene expression matrix of Hypothalamus}
\format{
A matrix with 14374 gene.
}
\source{
<Chen, Wu et al. 2017>
}
\usage{
gene_Hypothalamus
}
\description{
gene expression matrix of Hypothalamus contains 14373 gene ahd 727 cells
gene counts have been normalized.
}

