% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_index.R
\name{index_index}
\alias{index_index}
\title{Obtain clustering performance indicators}
\usage{
index_index(group.ed, label)
}
\arguments{
\item{group.ed}{Clustered labels}

\item{label}{True label}
}
\value{
a vector
}
\description{
Obtain clustering performance indicators
}
