% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_fusion.R
\name{network_fusion}
\alias{network_fusion}
\title{Similarity network fusion}
\usage{
network_fusion(data1, data2, K = 20, alpha = 0.5, T = 10)
}
\arguments{
\item{data1}{Cell similarity matrix}

\item{data2}{Another cell similarity matrix for network fusion}

\item{K}{Number of neighbors in K-nearest neighbors part of the algorithm.}

\item{alpha}{Variance for local model.}

\item{T}{Number of iterations for the diffusion process}
}
\value{
A consensus similarity matrix
}
\description{
The single-cell similarity network is integrated into the single-cell-cell similarity network based on the nonlinear combination method.
}
