\docType{data}
\name{PA_Hypothalamus}
\alias{PA_Hypothalamus}
\title{A demo:PA expression matrix of Hypothalamus}
\format{
A matrix with 7114 PACs.
}
\source{
<Chen, Wu et al. 2017>
}
\usage{
PA_Hypothalamus
}
\description{
PA expression matrix of Hypothalamus contains 7114 PACs ahd 727 cells
PAC counts have been normalized.
}
