% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAtoGE.R
\name{PAtoGE}
\alias{PAtoGE}
\title{Convert PA expression matrix to gene expression matrix}
\usage{
PAtoGE(data)
}
\arguments{
\item{data}{a PACdataset}
}
\value{
Gene expression matrix
}
\description{
Convert PA expression matrix to gene expression matrix
}
\examples{
PAtoRUD(data)
}
